auth: Log warning if rxkad-k5 is broken
We currently give no indication to the administrator if rxkad "keytab"
support is effectively disabled because libkrb5 initialization is
broken (krb5_init_context fails), or because we are not built with
rxkad "keytab" support. This causes rxkad-k5/rxkad-kdf connections to
fail, and without such logging it can be difficult to figure out why.
So, log a warning in these situations.

An initialization failure can happen if the local libkrb5 library is
misconfigured to such a degree that krb5_init_context fails. This is
known to occur on at least some Solaris 10 systems, where the default
system krb5.conf includes placeholders that are not valid
configuration.

This is a 1.6-only change. On master, this change would look very
different and is more difficult, and seems not worth the effort
because on master:

 - rxkad uses our local crypto functions where krb5_init_context
   cannot fail, not the system libkrb5
 - rxkad "keytab" support cannot be disabled
 - We do not have a function to initialize "keytab" support at
   startup, so it's less clear where such logs would go
 - We call krb5_init_context for every incoming connection, so a
   single failure at startup does not disable "keytab" support
   forever

Change-Id: I59a309f8c9546323651efa7490b2875db218a22f