libafs: put volume disk cache i/o in afs_UFSGetVolSlot
Move the reading of the volume items file to the afs_UFSGetVolSlot()
to make it more clear the volume items file is not accessed when
memcache is in effect.

This changes the afs_GetVolSlot to return an intialized volume slot,
if one can be gotten.

Reviewed-on: http://gerrit.openafs.org/10878
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: D Brashear <shadow@your-file-system.com>
(cherry picked from commit bbda24f9d324493e3f59c216a465fb2383d8f557)

Change-Id: Idcbaea71685542f9cc98b57a15cf2703890b759c