asetkey: allow adding keys without cell config for easier setup
Traditionally, when setting up a new cell, the bosserver would be
started in -noauth mode, which among other things, would create the
server configuration directory (i.e. /usr/afs/etc for Transarc
compatible paths) and a default CellServDB and ThisCell file.

The bosserver -noauth mode is obsolete and should no longer be used in
new setups. (kaserver is no longer used and pts now supports
-localauth.)  Instead, the service key should be added before the
bosserver is started for the first time, so the bosserver starts in the
normal authenticated mode.

However, asetkey add fails when the CellServDB and ThisCell cell
configuration files are not present, even though the cell configuration
is not need or used to add the service key.  The failure is a side
effect of the routine asetkey uses to open the server configuration to
modify the service keys.

Provide a new cell configuration routine to allow asetkey create, list,
and remove keys, without the requirement of the CellServDB and ThisCell
configuration.

Also, when adding a key, create the cell configuration directory, with
the same permissions as bosserver expects, if the configuration
directory is missing.

Usually the server configuration directory and the CellServDB and
ThisCell files are provided by packaging/installation tools.  However,
this change to asetkey simplifies the process when manually setting a
new cell.

Change-Id: I884f213430a5cfe0f51dc5339b40c943bd65afdd