libadmin: parse rxstat_* command line args with libcmd
Replace the custom command line parsing in the rxstat_* sample programs
with the common libcmd functions.

The command line syntax for the rxstat_* programs compatible with the
old custom command line parsing, but libcmd will make it easier to add
new options in the future.

This also fixes the incorrect usage messages in several of the commands.

Change-Id: I5df9b9a439bc12e7c1090abc9713e5245d666c1b