bozo: do not create client directories and symlinks
Traditionally, during startup the bosserver creates the client's
configuration directory and symlinks of the servers's CellServDB and
ThisCell in the client's directory.

Modern versions of the vos, pts, bos tools will use the server's
CellServDB and ThisCell files when invoked with -localauth, so the
client CellServDB and ThisCell files are only need if a client is
installed on the same host as the bosserver.

In commit 75d67780b42c1a7bfa506fcd230b28a6f293fcbd, the contributed
rpm packaging was updated to detect if the client's CellServDB and
ThisCell files were actually symlinks, and if so, remove the symlinks
and create regular files. This avoids clobbering the servers's
configuration when installing a client.

This commit removes the creation of the client files by the bosserver.
This will decouple the server from the client and help to avoid
overwriting the servers CellServDB configuration.

Change-Id: I8aba81185a52e6792d36ae28a430ebe1c23919d2