aklog: dump token data in with -d -d
As a debugging aid, print the token data just before setting the token
when aklog is invoked with two or more -d flags. Attempt to extract an
rxkad token and dump the fields of the rxkad token if possible.
Otherwise, print the raw hex dump of the opaque token data. Also print
the raw hexdump of the opaque token data when aklog is invoked with
three or more -d flags.

Change-Id: If76b26e9c7692554a3607de09dcd9b15addf12f3