redhat: fix conditional for kernel-debuginfo %files directive
Previously, commit 443dd5367e0cd9050ad39a6594c5be521271b4e9 added support
for a separate debuginfo package for the kernel module.  Unfortunately,
the %files directive for the kernel module debuginfo package was
incorrectly placed in the %if stanza of the build_userspace condition,
so the rpmbuild fails when attempting to build just the kernel module.

That is, when running rpmbuild with the options:

    rpmbuild --define "build_userspace 0" --define "build_modules 1" ...

rpmbuild fails with:

    RPM build errors:
       Installed (but unpackaged) file(s) found:
       /usr/lib/debug/lib/modules/.../extra/openafs/openafs.ko.debug

Fix this by moving the new %files directive out of the build_userspace
conditional.