libcmd: add exclusive option support
Sometimes command line options are mutually exclusive, for example, only
one of the fileserver options -S or -L may be given.  This commit adds a
feature to the libcmd syntax parser to check when more than one option
is given within a declared group of mutually exclusive command line
options.

This commit introduces option groups, which are used to identify which
options are mutually exclusive. A given syntax may have multiple option
groups.  All of the options within a group are not required options,
that is are set with the CMD_OPTIONAL flag.

Change-Id: I15a630b7e0b1fbc9f506cda39e787f1c307d5df0