viced: host,port -> sockaddr_in (interim)
This is a refactoring change to prepare for IPv6 support.  Convert the
host,port pairs to struct sockaddr_in in the host package.

Introduce a static function to compare two sockaddr_in values to improve
the code readability.

This patch intentionally uses the struct sockaddr_in and pointers to
struct sockaddr_in instead of attempting to pass sockaddrs with more
generic struct sockaddr * type, which would require type casting.

Change-Id: I1b7bcec2f89017ce621c19cf568b6b0aac257647