vos: return true when GetServerAndPart finds a site
Change the GetServerAndPart() function and callers to use a return value
to indicate when a volume site is found, instead checking for special
values of the output arguments.

While here, also add some doxygen function comment header and fix some
typos in the comments within the function.  This is a refactoring change
to clean up the code in preparation for other changes and should not
introduce any functional change.

Change-Id: I10b43054b1bf9e6757ccdc95cb4559ab8b6dc013