vos: Return true when GetServerAndPart finds a site
Change the GetServerAndPart() function to return true when a volume site
in the vldb entry is found. Do not change the output arguments unless
the site is found.

Also add a function comment header and fix some comment typos.  This is
a refactoring change in preparation of supporting non-integer server
ids.

Change-Id: I10b43054b1bf9e6757ccdc95cb4559ab8b6dc013