volser: Standardize error returns in vos.c
Many functions within vos.c have returns embedded within the body of the
function.

An existing local macro, ERROR_EXIT, was available that sets a variable
to set a value and branches to a common exit point.

Use ERROR_EXIT to replace the embedded returns where it makes sense to
do so (typically within the functions processing the command line
parameters).

Change-Id: I73910d97d29ba4bcd31557199328764de96eaff3