vos: vos remove -partition requires a -server option
As documented in the vos remove man page, the vos remove -server option
should also be given when the -partition option is specified.
Currently, when the vos remove is run with the -partition option without
the -server option, the partition id verification fails since the server
value is not set.  Instead, give a message the user to indicate they
should give the -server option as well, then exit.

Use the same error message text as is already used for other commands.

Change-Id: I6b878678e28b34250e63d2d082747f6fd416972d