volser: Deprecate MapPartIdIntoName
The old MapPartIdIntoName() function is an old implementation of the
routine to generate a partition name from number.  This old
implementation is missing error checking and is generally unsafe since
it assumes the caller allocates enough space for the destination string.

Since this is an exported function provide a wrapper using the
volutil_PartitionName_r() function.

Change-Id: I23f58282fde5f7b86af1b0fa42d1027f358f99e6