volser: deprecate MapPartIdIntoName
The old MapPartIdIntoName() function is an old implementation of the
routine to generate a partition name from number.  This old
implementation is missing error checking and is generally unsafe since
it assumes the caller allocates enough space for the destination string.

Update vos.c and vsproc.c to use the volutil routines to get partition
names instead of MapPartIdIntoName(). Provide a wrapper for any other
users of libvolser.