volser: Cleanup server and partition name formatting
Introduce the servid2str() and partid2str() helper functions to get
string representations of server and parititions for human readable
output. Define the serverName and partName structures so we have
consistent buffer sizes for the output strings.

The servid2str() function is a wrapper of the ternary operator
to print the dotted decimal address when running in noresolve mode
or to lookup the hostname otherwise.

The partid2str() function is a wrapper of the volume utility
volutil_PartitionName_r() to get the partition name from an id.

Update all of the printf arguments in vsprocs.c and vos.c to use these
new helpers and remove the old MapPartIntoName() calls before the
printfs. Fix the whitespace errors in the printfs which were introduced
in commit 40d41c191bdb64256d4ccf6a5f1bab8124249e63
(vos-noresolve-option-20080410).

Use the volutil_PartitionName2_r() in the vos.c functions which need to
get the partition names for processing, and honor the error code from
volutil_PartitionName2_r() in those code paths.

Change-Id: I1deff45e8eb325cca7042eb2d617c0e7dc29a5be