auth: fix bosserver crash when creating default cell config files
Commit 93b26c6f55245e2187e574eb928f5e0ce66a245e added the cellservDB
field to the afsconf_dir structure to track the CellServDB pathname.
This commit also changed the afsconf_SetCellInfo() and
afsconf_SetExtendedCellInfo() functions to use the new cellservDB member
to open the CellServDB file.

Unfortunately, the bosserver intentionally calls afsconf_SetCellInfo()
with a NULL afsconf_dir pointer when attempting to create the default
CellServDB and ThisCell files (e.g., "localcell"), which causes the
bosserver to crash on startup when the cell configuration is not present.

To fix this, change the afsconf_SetCellInfo() wrapper to make a
temporary afsconf_dir object with a valid cellservDB value and pass
that to the implementation in afsconf_SetExtendedCellInfo().

From the comments above afsconf_SetCellInfo(), it appears the
afsconf_dir argument was added in order to refresh the configuration
cache when updating cell information with bosserver RPCs.  Since the
these functions now deal with the afsconf_dir data structure, and now we
need to access static methods in cellconfig.c to get the cellservDB
value, move the afsconf_SetCellInfo(), afsconf_SetExtendedCellInfo(),
and its static helper function, from writeconfig.c to cellconfig.c, and
retire the writeconfig.c file.

Also, change the signature of the _afsconf_CellServDBPath() static
function to take just the base directory name of the CellServDB file
instead of the entire afsconf_dir data object. This makes it more clear
which fields are needed to compose the path to the CellServDB.

Change-Id: I8d36f7c8afe6b4e13bfd04c421bf1109d1eb4238