util: Prevent server log deadlocks with pthread_atfork
The salvageserver creates several worker threads then enters a loop to
fork child processes to perform salvages. If the fork happens as a child
process has the server logging mutex locked, the child process inherits
a locked mutex which will never be unlocked and deadlocks, since the
child process does not

To prevent this deadlock, register handlers with pthread_atfork() so the
server logging mutex will be held during the fork. The fork will be
blocked until the worker thread releases the server lock, and the lock
will be held until the fork is complete, and the lock will be released
in the parent and child threads.

Change-Id: Ib105b026b347e792ce82e3c92723aee724cfd153