cmd: Allow "--" as an end-of-options delimiter
Currently, it's impossible to give any libcmd option a value that
starts with a hyphen followed by a non-digit. For example, if you
wanted to lookup a volume with the name "-foo", you can't give that
name directly to vos, since libcmd interprets it as a (nonexistent)
option name:

    $ vos listvldb -foo
    vos: Unrecognized or ambiguous switch '-foo'; type 'vos help listvldb' for detailed help
    $ vos listvldb -name -foo
    vos: Unrecognized or ambiguous switch '-foo'; type 'vos help listvldb' for detailed help

To fix this, catch up to POSIX.2 from 1992 by interpreting "--" as a
delimiter for the end of options; anything after the "--" is
interpreted as an operand. This makes it possible to run a command
like the above example:

    $ vos listvldb -- -foo
    VLDB: no such entry

Since now specifying an option (or alias) name of "--" won't work
properly, prevent the creation of options with such a name. Add the
new error code CMD_BADNAME to handle this case.

Change-Id: I9670a1fe398400ec1b806704a7877185accd564e