[phase1] [server] bozo: Allocate RPC output strings with strdup()
Instead of allocating 256 byte buffers and then filling them, use
strdup() to allocate the RPC output strings which report information
about bnodes.

The allocated buffers are currently too small by one character and
pre-allocating the strings prevents us from significantly increasing the
size of the output strings in the future. (The xdr string size of 256 is
the length of the maximum string, not including the terminating nul.)

This commit updates how we create the output strings for the following
RPCS:

SBOZO_EnumerateInstance
SBOZO_GetInstanceInfo
SBOZO_GetInstanceParm
SBOZO_GetStatus

The other bosserver RPCs already used strdup() to allocate and populate
RPC output strings, except for SBOZO_ListSUsers(), which returns
usernames.

This commit changes the bnode_GetParm() and bnode_GetString() to return
a copy of the string, instead of filling a buffer. The caller must free
the allocated copy.

Change-Id: I1fd4f2a83214fcca73ae0939f1a27d698fdff410