[phase1] [prep] bozo: Move file renames and global init out of ReadBozoFile()
Move the code to rename the BozoInit and BosConfig.new files,
and the code to initialize the bosserver restart and ne out
of the ReadBozoFile() function and to the bosserver main() function.

These file renames only happen on bosserver startup so they are better
suited to be in the main() function, and moving them out of
ReadBozoFile() makes it easier to create unit tests for that function in
future commits.

wip: move ktime init out of read

Change-Id: Ib04171eee3b0b0fd2b638111e95ccbe807ce83f5