[phase1] [prep] bozo: Remove bozo_fileName global
Currently, the bozo_fileName global is set to a copy of the BosConfig
file path and is never changed or freed. A zero is always passed to
ReadBozoFile() and WriteBozoFile(), and the bozo_fileName global value
is used to open the file.

Clean this up by removing the unncessary bozo_fileName global. Change
ReadBozoFile() and WriteBozoFile() to expect a non-NULL argument which
is the file path of the configuration file to read/write, and update the
call sites to specify the dirname symbol for the BosConfig file path.

Note that we keep the file path argument in ReadBozoFile() and
WriteBozoFile() to make it easier to write unit tests for those
functions in a future commit.

Change-Id: I8cdcadbcedbc7e7c0d4d3921724954205d55369c