tests: Add BosConfig tests
Add c-tap unit tests for the ReadBozoFile() and WriteBozoFile() functions.

This commit adds a "mock" bnode for the purposes testing the reading and
writing of the test BosConfig files.  The mock bnode just has the
contents of the bnode parms and the instance name and is registered as
the normal bosserver bnode types.

ReadBozoFile() is tested by creating a test BosConfig file and then call
ReadBozoFile() to read the test file and verify the number and contents
of the bnodes match what is expected. ReadBozoFile() calls
bnode_Create() to create the mock bnodes.

WriteBozoFile() is tested by first creating a set of mock bnodes, then
writing them to a test BosConfig file. WriteBozoFile() calls
bnode_GetParm() on the mock bnodes to get the parm strings to be
written.

Change-Id: I9085b379c05f9b0e4ec847d667cc9c73a081e04d