[phase3] [cleanup] config: Replace custom line parsing with with strtok()
The `config` program is a custom build tool we use to generate Makefiles
from MakefileProto files. `config` provides a custom templating syntax
to include or exclude lines depending on the target sysname.

MakefileProto directives are currently parsed by a modified
"ParseLine()" function, which was copied to various locations in the
OpenAFS code base.

To clean this up, replace the custom parsing code with the standard
strtok() string utility to split the words enclosed within the '<' and
'>' angle brackets in the MakefileProto templates.

Do not treat the start '<' and ending '>' angle brackets as whitespace.
Instead use them to enclose the list of names to be matched.

Add a line number count for error messages.

Change-Id: I5b4057fa808e4be695ce49521fc26ab8640002ad