[phase3] [cleanup] libadmin: Remove duplicated ktime_ParsePeriodic()
The libadmin developers copied the libutil ktime_ParsePeriodic()
function and associated functions from libutil to a test program in
libadmin, then renamed the struct ktime to BosRestartTime_t to match the
name used in libadmin.

Remove the duplicated code in libadmin and just call the libutil
ktime_ParsePeriodic() function.

Fortunately, the libadmin test program already links against libutil, so
all that is needed is to add the additional headers and remove the
duplicated code. We keep the "day" names array, since it is used by the
test program.

This commit intended remove yet another copy of the old "ParseLine"
function which was duplicated in various places in the OpenAFS code
base.

Change-Id: I96d63eab48da489116defff204d0f5c7f2294be1