bozo: Remove duplicate code to check if restarts are required
The code to stat the program file to determine if a bosserver managed
process should be restarted was duplicated in five places.

Add a new common function called bnode_IsRestartRequired() to parse the
command line, stat the program file, and check to see if the ctime was
changed since the last restart.  Update the simple and fs/dafs bnode
types to use the new function instead of duplicating code.

Change-Id: Ifa88426e6a02eef0be5a510c751c06d6a09717a7