cmd: Add cmd_Split() and cmd_Join() to split and join command lines
cmd_Split() will split a shell-like command line into an argument vector
suitable for exec*()-style functions, without artificial limits on the
size of arguments or the number of arguments.  cmd_Split() supports
argument quoting and nested quoting, so whitespace can be included in
arguments.  Both single quotes (') and double quotes (") are supported.
For example, the string:

  "don't worry" be happy

is split into the three strings:

  don't worry
  be
  happy

cmd_Split() also supports backslash ('\') escaping on unix-like
platforms.  Backslash '\' escaping is not supported on Windows, as in
conventional, to avoid issues with path separators on that platform.

The cmd_FreeSplit() may be used to free the resulting cmd_Split()
argument vector.  (Note the existing cmd_FreeArgv() only frees the
strings of an argv vector.)

cmd_Join() is the complement of cmd_Split(). The cmd_Join() function
will combine the strings in an argument vector to create a properly
quoted shell-like command string, without artificial limits on the size
of the arguments or the number of arguments.

This commit includes a set of c-tap tests to put these new functions
through their paces.

Change-Id: I7eb3771aceb81dd91b1a7235c6c03a382815bdd7