[phase3] [cleanup] cmd: Rework cmd_ParseLine()
The cmd_ParseLine() function splits a command argument string into an
argument string vector with a placeholder at argv[0]. The implementation
is based on the ParseLine function, which was copied to several places
in the code tree, with changes to attempt to support quoting.

This commit reimplements cmd_ParseLine() with the same code used by
cmd_Spit() to overcome several limitations of the current
cmd_ParseLine():

* The token buffer is fixed length, on the stack.
* The number of arguments is limited by the size of the caller provided buffer.
* Nested quotes are not supported, e.g., "don't worry"
* Missing quotes are ignored.
* Escaping with '\' is not supported on unix-like platforms

The function signature has not been changed in order to maintain
compatibility with existing callers.

The cmd_FreeArgv() function was updated to clear arguments after freeing
them to avoid double free errors.

Change-Id: I6eeeb0bf8ffbc30978e45e3f49d559cde7bb7b6e