[Do not merge] tests: Run auth superuser tests in the foreground
NOTE: This is not ready for upstream. The current method is better at
cleaning up since the main process can kill the client and server
processes. When running the client in the foreground, it can crash and
the atexit handler is not run, so we have an orhaned server process.  If
we want this in upstream, I think running the client in the foreground
should be a run time option.  For now, just keep this as a developer
only commit in my sandbox.

The superuser-t tests fork a client and server process. The actual
tests are in the client process, the parent process just waits for
the client to finish and then cleans up by killing the child processes
and removing the test configuration files (a temporary UserList file).

To make it easier to run the tests in a debugger, just fork the server
process and run the client checks in the foreground.  Register a c-tap
cleanup handlers to kill the server process and remove the test config.

Print extra messages to stderr when the C_TAP_VERBOSE env var is set.

Change-Id: Ia8e22927be4d8267fdbd95e2cc89d521e94f9f28