tests: Run auth superuser tests in the foreground
The superuser-t tests fork a client and server process. The actual
tests are in the client process, the parent process just waits for
the client to finish and then cleans up by killing the child processes
and removing the test configuration files (a temporary UserList file).

To make it easier to run the tests in a debugger, just fork the server
process and run the client checks in the foreground.  Register a c-tap
cleanup handlers to kill the server process and remove the test config.

Print extra messages to stderr when the C_TAP_VERBOSE env var is set.

Change-Id: Ia8e22927be4d8267fdbd95e2cc89d521e94f9f28