auth: Do not truncate UserList entries
In OpenAFS 1.0, only local cell users where supported in the UserList
file, so it was assumed the maximum entry size was 63 (MAXKTCNAMELEN -
1) characters.  The bos adduser, removeuser, and listusers commands have
always supported longer names, ostensibly up to 256 (BOZO_BSSIZE)
characters, and of course, although not recommended, users with root
access can manually edit the UserList file.

Unfortunately, due to an incorrect sscanf format specifier in the original
code, UserList lines longer than 63 characters caused a buffer overflow
by one character.  That is, the code to parse UserList lines was like:

   char name[64];                /* correct (MAXKTCNAMELEN) */
   sscanf(buffer, "%64s", name); /* should have been %63s */

In addition, this code suffers from a missing truncation check, so only
the first 64 characters of a UserList entry are signficant during
superuser checks and when removing entries with bos.

The overflow was fixed in commit 61cf9fa731654b943a271dcf585b95062fbe7a4c
(circa 2003), however, that fix increased the buffer size by one byte instead
of correcting the sscanf() specifier. That is, the code was changed to:

   char name[64 + 1];
   sscanf(buffer, "%64s", name);

instead of:

   char name[64];
   sscanf(buffer, "%63s", name);

The missing truncation check was not addressed as well.

Commit e69d27e8f0169f1e792887305b130e5fd2a7a4af added support for fully
qualified names, that is, names in the form <name>.<instance>@<REALM>.
After that change fully names extracted from Rx calls may be longer than
63 characters, but due to the silent truncation, only the first 64
characters of the UserList entries are signficant in superuser checks.

Fix the UserList name truncation by removing the unneeded sscanf. This
makes the full UserList entry significant for superuser checks and when
removing entries with bos removeuser.

Add code to trim trailing newlines which were removed by sscanf before
this change.  The line is intentionally trimmed before parsing both old
and new style identity formats. This removes an unwanted newlines from
the display name in new style identities as well, which seems to be an
oversight when new style identities were added.

Change-Id: Ifa5366259cd178f4e68d76fc913ffc31dd937abe