auth: Separate string buffer for UserList filename
Ever since OpenAFS 1.0, the functions in userok.c with read the UserList
file have a single buffer which does double duty. The buffer is first
filled with the fully qualified filename of the UserList file to open
the file, then after the file has been opened, the same buffer is reused
to read the contents of the UserList file, line by line.

Unfortunately, the size of the buffer is tied to the max path size,
which is platform dependent and the UserList file format is platform
independent.  Instead, create a separate buffer for the filename and use
asprint() to dynamically compose the fully qualified path name of the
UserList file.

Change-Id: I91add619a0a1defc6cbe88319bcf5bc3c368a262