auth: Platform independent UserList max line size
Currently, the maximum line size of the UserList file is platform
dependent, because we reused the filename buffer (which was tied to the
max path size) for the line buffer.

In this change, the UserList line buffer size is set to a platform
independent value of 2048. Positions are needed for the newline, the
terminating nul character, and one character for a truncation check, so
the max line contents is 2045 characters.

Also add a sanity check when adding new entries to the UserList to avoid
creating lines longer that the maximum buffer size.  Currently, the
bosserver is limited to 256 (BOZO_BSSIZE) number of charaters, but that
limitation may be removed in the future.

Note: In the future, it may be better to use the standard getline()
function to read lines instead of imposing a predetermined maximum line
size, however currently the code paths used by the fileserver read the
UserList with the custom Bufio* functions. These functions seem to be in
place to workaound the limited number of file descriptors available in
very old platforms (such as 32-bit Solaris 10, according to comments in
the tree).

Change-Id: Ic37ab2bea2de1808a51762f871a73398a56eb109